+++ src/interfaces/statusCombined.ts
@@ -1,7 +1,0 @@
-import { StatusItem } from "../lib/statusBar";
-
-export interface StatusCombined {
-  cpu: StatusItem;
-  ram: StatusItem;
-  battery: StatusItem;
-}